pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "century_ring"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
readme = "README.rst"
license = "GPL-3.0-or-later"
version = "25.0.0"
dependencies = ["attrs>=24.3.0"]

[project.optional-dependencies]
async = [
    "anyio>=4.8.0",
    "sniffio>=1.3.1",
]
trio = [
    "trio>=0.28.0",
    "century_ring[async]"
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
    "pyright>=1.1.392.post0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "sphinx>=8.1.3",
    "esbonio>=0.16.5",
    "sphinx-autodoc-typehints>=3.0.1",
    "sphinx-rtd-theme>=3.0.2",
]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "pysrc"
module-name = "century_ring._century_ring"

[tool.pyright]
pythonVersion = "3.12"
include = ["pysrc"]
typeCheckingMode = "strict"

reportImplicitOverride = "warning"
reportUnusedVariable = false
reportIncompatibleVariableOverride = false
reportPrivateUsage = false

[tool.ruff]
target-version = "py312"
respect-gitignore = true
# fix = true
src = ["pysrc/"]
line-length = 100
output-format = "full"

[tool.ruff.lint]
select = [
    "RUF",
    "F",
    "E",
    "W",
    "I",
    "UP",
    "B",
    "Q",
    "YTT",
    "PYI",
    "SIM",
    "RET",
    "PERF",
    "PIE",
    "SIM",
    "C4",

    # specific docstring rules
    "D201", "D204", "D211", "D419",   # less ugly docstrings
]
ignore = [
    "W291",
    "W293",
    "PYI021",
]

[tool.ruff.lint.per-file-ignores]
"tests/test_*" = ["D"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.esbonio.sphinx]
#pythonCommand = ["/home/lura/dev/libs/century-ring/.venv/bin/python3"]
pythonCommand = ["${venv:/home/lura/dev/libs/century-ring/.venv}"]
